Further code cleaning
Former-commit-id: d01966457737ee4580902c6cf53f05dfcc6923c7

+++ src/librapid/math/vector.hpp
@@ -4,7 +4,7 @@ namespace librapid {
 	template<typename Scalar, i64 Dims>
 	using Vec = VecImpl<Scalar, Dims, Vc::SimdArray<Scalar, Dims>>;
 
-	template<typename Scalar, i32 Dims, typename StorageType>
+	template<typename Scalar, i64 Dims, typename StorageType>
 	class VecImpl {
 	public:
 		using Mask = Vc::Mask<Scalar, struct Vc::simd_abi::fixed_size<Dims>>;
